一時的にミサイルを無効化。out/をcommit

+++ src/Base.java
@@ -231,7 +231,7 @@ public class Base implements GameObject {
 	}
 
 	@Override
-	public double getHP() {
+	public int getHP() {
 		return hp;
 	}
 }
